Create class Enemy and make adjustments to the engine
Now the game runs, map is drawn, one enemy is drawn, and the
enemy moves through the hardcoded path of the map. The enemy
gets deleted when it hits the last waypoint of the path.

Also added:

* Images for one enemy's sprite

* Map image

* Debug drawing methods (e.g. bounding boxes and sprite origins)

Nevertheless, this commit's diff is a bit messy.

The adjustments made to the game engine are large scale.
The engine, too, is in development phase and thus needs
modifications all the time.